+++ readium/streamer/CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 **Warning:** Features marked as *alpha* may change or be removed in a future release without notice. Use with caution.
 
-## [Unreleased]
+<!--## [Unreleased]-->
+
+## [2.1.0]
 
 ### Added
 
@@ -14,7 +16,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-* Upgraded to Kotlin 1.5.21 and Gradle 7.1.1
+* Upgraded to Kotlin 1.5.31 and Gradle 7.1.1
 * The default EPUB positions service now uses the archive entry length when available. [This is similar to how Adobe RMSDK generates page numbers](https://github.com/readium/architecture/issues/123).
     * To use the former strategy, create the `Streamer` with: `Streamer(parsers = listOf(EpubParser(reflowablePositionsStrategy = OriginalLength(pageLength = 1024))))`
 
@@ -110,4 +112,5 @@ server.loadCustomResource(assets.open("scripts/highlight.js"), "highlight.js", I
 [2.0.0-beta.1]: https://github.com/readium/r2-streamer-kotlin/compare/2.0.0-alpha.2...2.0.0-beta.1
 [2.0.0-beta.2]: https://github.com/readium/r2-streamer-kotlin/compare/2.0.0-beta.1...2.0.0-beta.2
 [2.0.0]: https://github.com/readium/r2-streamer-kotlin/compare/2.0.0-beta.2...2.0.0
+[2.1.0]: https://github.com/readium/r2-streamer-kotlin/compare/2.0.0...2.1.0